Refactored utilites for command.py new file
